requirements.txt
anthropic>=0.39.0
python-dotenv>=1.0.0
click>=8.1.0
pytest>=8.0.0
